requirements.txt
alembic==1.8.0
appnope==0.1.3
asttokens==2.0.5
Authlib==1.0.1
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.2
black==22.3.0
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
executing==0.8.3
Flask==2.1.2
Flask-Admin==1.6.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
ipdb==0.13.9
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
oauthlib==3.2.0
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prompt-toolkit==3.0.29
psycopg-binary==3.0.15
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.37
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
